requirements.txt
fastapi
python-dotenv
uvicorn
httpx
gunicorn
ollama